+++ shape/cylinder.c
@@ -2,6 +2,6 @@
 
 #include "abstract_shape.h"
 
-void cylinder(int edges, double radius, double height, struct vertex** vertices, long* vertices_size, GLushort** indices, long* indices_size) {
+void cylinder(int edges, float radius, float height, struct vertex** vertices, long* vertices_size, GLushort** indices, long* indices_size) {
   abstract_shape(edges, radius, height, 0, 0, vertices, vertices_size, indices, indices_size);
 }
