+++ shape/polygon.c
@@ -2,6 +2,6 @@
 
 #include "abstract_shape.h"
 
-void polygon(int edges, double radius, struct vertex** vertices, long* vertices_size, GLushort** indices, long* indices_size) {
+void polygon(int edges, float radius, struct vertex** vertices, long* vertices_size, GLushort** indices, long* indices_size) {
   abstract_shape(edges, radius, 0, 0, 0, vertices, vertices_size, indices, indices_size);
 }
